pyproject.toml
[project]
name = "nlu-assignment-2"
version = "0.1.0"
description = "NLU assignment: corpus + Word2Vec experiments"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12",
    "gensim>=4.3",
    "langdetect>=1.0.9",
    "lxml>=5.0",
    "matplotlib>=3.8",
    "nltk>=3.8",
    "pdfplumber>=0.11",
    "requests>=2.31",
    "scikit-learn>=1.3",
    "wordcloud>=1.9",
]

[project.optional-dependencies]
problem2 = [
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "torch>=2.0",
]

[tool.uv]
package = false
